worker/requirements.txt
supabase
requests
pymupdf
pyyaml
